feat: Don't announce that no jester exists if jester is disabled
Before this change and if the jester was installed but disabled, every
round start a mstack message was printed informing all players that there
is no jester in this round.

This message is unneeded under the assumption that every player knows
that the jester is disabled.
While this assumption might not always be true, i think the possible
annoyance this message brings trumps the information given.